requirements.txt
Django==3.2.8
djangorestframework==3.12.4
feedparser==6.0.8
psycopg==2.9.2
celery==5.2.1
